pyproject.toml
[tool.poetry]
name = "fixup-md-img"
version = "0.1.0"
description = ""
authors = ["Patrice Ferlet <patrice.ferlet@smile.fr>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
